terraform/external_processor/module/lambda/requirements.txt
requests
boto3
psycopg2-binary